requirements.txt
Django==4.0.1
django-rest-framework==0.1.0
psycopg2==2.9.3
PyJWT==2.3.0
